+++ src/Symfony/Component/Translation/Bridge/PoEditor/Provider/PoEditorProviderFactory.php
@@ -49,7 +49,7 @@ final class PoEditorProviderFactory extends AbstractProviderFactory
     {
         if (self::SCHEME === $dsn->getScheme()) {
             $client = $this->client->withOptions([
-                'base_uri' => sprintf('%s%s', 'default' === $dsn->getHost() ? self::HOST : $dsn->getHost(), $dsn->getPort() ? ':' . $dsn->getPort() : ''),
+                'base_uri' => sprintf('https://%s%s', 'default' === $dsn->getHost() ? self::HOST : $dsn->getHost(), $dsn->getPort() ? ':' . $dsn->getPort() : ''),
             ]);
 
             return new PoEditorProvider($this->getUser($dsn), $this->getPassword($dsn), $client, $this->loader, $this->logger, $this->defaultLocale);